+++ lib/generators/phrase_ota/templates/README.md
