+++ lib/generators/phrase_ota/templates/phrase_ota.rb
